renamed block production directory
