+++ src/chain/factory/block/body.ts
@@ -2,9 +2,9 @@
  * @module chain/blockAssembly
  */
 
-import {OpPool} from "../../opPool";
-import {BeaconBlockBody, BeaconState, bytes96} from "../../types";
-import {MAX_ATTESTATIONS, MAX_ATTESTER_SLASHINGS, MAX_PROPOSER_SLASHINGS, MAX_VOLUNTARY_EXITS} from "../../constants";
+import {OpPool} from "../../../opPool";
+import {BeaconBlockBody, BeaconState, bytes96} from "../../../types";
+import {MAX_ATTESTATIONS, MAX_ATTESTER_SLASHINGS, MAX_PROPOSER_SLASHINGS, MAX_VOLUNTARY_EXITS} from "../../../constants";
 import {bestVoteData} from "./eth1Data";
 
 export async function assembleBody(
