+++ src/chain/factory/block/deposits.ts
@@ -2,8 +2,8 @@
  * @module chain/blockAssembly
  */
 
-import {BeaconState, Deposit} from "../../types";
-import {OpPool} from "../../opPool";
+import {BeaconState, Deposit} from "../../../types";
+import {OpPool} from "../../../opPool";
 
 export async function blockDeposits(opPool: OpPool, state: BeaconState): Promise<Deposit[]> {
   if(state.latestEth1Data.depositCount > state.depositIndex) {
